// src/interfaces.ts
export interface Field{
    name: string
    street: string
    id: number
    row: number
    width: number
    height: number
    orientation: string
    boardNumber: number
    isProperty: boolean
    propertyColor: string
    isTrain: boolean
    isUtility: boolean
    isChance: boolean
    isKanzlei: boolean
    isEvent: boolean
    isGo: boolean
    isJail: boolean
    isParking: boolean
    isPopo: boolean
    isSpacer: boolean
    icon: string
}